requirements.txt
anyio==3.5.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
ecdsa==0.17.0
fastapi==0.72.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.9
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0
